Debug and assign correct index names
This code works without any bugs and it successfuly scales the models.

The required inputs are: AJ's .osim file and the ManualScaleConfiguration.xml file (in the "Simulations/IK" directory).

The outputs are: 14 files of scaling configurations in .xml format. At the moment this code overwrites the .osim file instead of creating 14 different .osim files.  This will be fixed in the next version.