Add IK loop to the batch model scaling code
This code successfuly runs IK for the 14 scaled .osim files. At the moment the code overwrites the output file, therefore the next step is to alter this part of the code (I think it is lines 153, 157 and all related 'output' files) to record IK results for all the models without overwriting.